// src/types/valueName.ts
export enum ValueName {
    Honesty = 'honesty',
    Freedom = 'freedom',
    Power = 'power',
    Compassion = 'compassion',
    Justice = 'justice',
    Obedience = 'obedience',
    Tolerance = 'tolerance',
    Achievement = 'achievement',
    Loyalty = 'loyalty',
    SelfExpression = 'self_expression',
    Responsibility = 'responsibility',
    Mercy = 'mercy',
    Equality = 'equality',
    Friendship = 'friendship',
    Leadership = 'leadership',
    Order = 'order',
    Faithfulness = 'faithfulness',
    Wisdom = 'wisdom',
    Security = 'security',
    Tradition = 'tradition',
    Ambition = 'ambition',
    Prestige = 'prestige',
    Honor = 'honor',
    Harmony = 'harmony',
    Curiosity = 'curiosity',
    Pleasure = 'pleasure',
    Humility = 'humility',
    Independence = 'independence'
  }
  
  export const ValueNameLabelMap: Record<ValueName, string> = {
    [ValueName.Honesty]: 'Честность',
    [ValueName.Freedom]: 'Свобода',
    [ValueName.Power]: 'Сила',
    [ValueName.Compassion]: 'Сострадание',
    [ValueName.Justice]: 'Справедливость',
    [ValueName.Obedience]: 'Послушание',
    [ValueName.Tolerance]: 'Терпимость',
    [ValueName.Achievement]: 'Достижение',
    [ValueName.Loyalty]: 'Преданность',
    [ValueName.SelfExpression]: 'Самовыражение',
    [ValueName.Responsibility]: 'Ответственность',
    [ValueName.Mercy]: 'Милосердие',
    [ValueName.Equality]: 'Равенство',
    [ValueName.Friendship]: 'Дружба',
    [ValueName.Leadership]: 'Лидерство',
    [ValueName.Order]: 'Порядок',
    [ValueName.Faithfulness]: 'Верность',
    [ValueName.Wisdom]: 'Мудрость',
    [ValueName.Security]: 'Безопасность',
    [ValueName.Tradition]: 'Традиции',
    [ValueName.Ambition]: 'Амбиции',
    [ValueName.Prestige]: 'Престиж',
    [ValueName.Honor]: 'Честь',
    [ValueName.Harmony]: 'Гармония',
    [ValueName.Curiosity]: 'Любознательность',
    [ValueName.Pleasure]: 'Удовольствие',
    [ValueName.Humility]: 'Смирение',
    [ValueName.Independence]: 'Независимость'
  };
  